feat: landing fixes, hero/services sections
