+++ src/app/[locale]/globals.css
@@ -14,6 +14,8 @@
   /*Fuentes*/
   --montserrat: "Montserrat", serif;
   --dm-sans: "DM Sans", serif;
+
+  scroll-behavior: smooth;
 }
 
 body {
