+++ src/app/[locale]/page.js
@@ -1,15 +1,12 @@
-"use client"
-import "./globals.css"
-import MainPage from "./components/MainPage/MainPage.js"
-
- 
+import "./globals.css";
+import MainPage from "./components/MainPage/MainPage";
 
 export default function Home() {
-    return (<>
-      
-    <main>
-      <MainPage/>     
-    </main>
+  return (
+    <>
+      <main>
+        <MainPage />
+      </main>
     </>
-  )
+  );
 }
