+++ src/messages/en.json
@@ -10,7 +10,8 @@
     "data-content-11": "of forest",
     "data-content-2": "soccer",
     "data-content-21": "stadiums",
-    "data-content-3": "offset",
+    "data-content-3": "tn CO₂",
+    "data-content-31": "offset",
     "data-content-1-mobile": "hectares of forest",
     "data-content-2-mobile": "soccer stadiums",
     "data-content-3-mobile": "tn CO2 offset"
@@ -20,10 +21,7 @@
     "title": "SERVICES",
     "subtitle-1": "PRESERVE",
     "subtitle-2": "NATIVE TERRITORIES",
-    "text-1": "Native forests are one of the main producers",
-    "text-2": " of oxygen, they are essential for us and",
-    "text-3": " all living beings on the planet. For this reason, we must",
-    "text-4": "protect native territories from being deforested.",
+    "text": "Native forests are one of the main producers of oxygen, they are essential for us and all living beings on the planet. For this reason, we must protect native territories from being deforested.",
     "service-card-1-title": "Measure",
     "service-card-1-text": "Know your environmental impact. Use the calculator and find out your CO2 emissions.",
     "service-card-2-title": "Compensate",
