+++ src/messages/es.json
@@ -10,7 +10,8 @@
     "data-content-11": "de bosque",
     "data-content-2": "estadios",
     "data-content-21": "de fútbol",
-    "data-content-3": "compensadas",
+    "data-content-3": "tn CO₂",
+    "data-content-31": "compensadas",
     "data-content-1-mobile": "hectareas de bosque",
     "data-content-2-mobile": "estadios de fútbol",
     "data-content-3-mobile": "tn CO2 compensadas"
@@ -20,10 +21,7 @@
     "title": "SERVICIOS",
     "subtitle-1": "CONSERVÁ LOS",
     "subtitle-2": "TERRITORIOS NATIVOS",
-    "text-1": "Los bosques nativos son uno de los principales",
-    "text-2": "productores de oxígeno, son esenciales para nosotros y",
-    "text-3": "todos los seres vivos del planeta. Por esto, debemos",
-    "text-4": "proteger los territorios nativos de ser deforestados.",
+    "text": "Los bosques nativos son uno de los principales productores de oxígeno, son esenciales para nosotros y todos los seres vivos del planeta. Por esto, debemos proteger los territorios nativos de ser deforestados.",
     "service-card-1-title": "Medí",
     "service-card-1-text": "Conocé tu impacto ambiental. Utilizá la calculadora y descubrí tus emisiones de CO2",
     "service-card-2-title": "Compensá",
